Separating swagger related functions in a module.
