+++ lib/index.js
@@ -9,6 +9,7 @@ var path = require('path');
 var doctrine = require('doctrine');
 var jsYaml = require('js-yaml');
 var parser = require('swagger-parser');
+var swaggerHelpers = require('./swagger-helpers');
 
 /**
  * Parses the provided API file for JSDoc comments.
@@ -42,7 +43,6 @@ function parseApiFile(file) {
   };
 }
 
-
 /**
  * Filters JSDoc comments for those tagged with '@swagger'
  * @function
@@ -64,115 +64,6 @@ function filterJsDocComments(jsDocComments) {
   }
 
   return swaggerJsDocComments;
-}
-
-/**
- * Merges two objects
- * @function
- * @param {object} obj1 - Object 1
- * @param {object} obj2 - Object 2
- * @returns {object} Merged Object
- */
-function objectMerge(obj1, obj2) {
-  var obj3 = {};
-  for (var attr in obj1) {
-    if (obj1.hasOwnProperty(attr)) {
-      obj3[attr] = obj1[attr];
-    }
-  }
-  for (var name in obj2) {
-    if (obj2.hasOwnProperty(name)) {
-      obj3[name] = obj2[name];
-    }
-  }
-  return obj3;
-}
-
-/**
- * Yields a warning for a given deprecated property.
- * @function
- * @param {string} propertyName - The property to warn about.
- */
-function deprecatedPropertyWarning(propertyName) {
-  if (propertyName === 'tag') {
-    console.warn('tag will be deprecated in v2.0.0');
-    console.warn('Please use tags as it aligns with the swagger v2.0 spec.');
-  }
-}
-
-/**
- * Adds the tags property to a swagger object.
- * @function
- * @param {object} conf - Flexible configuration.
- */
-function attachTags(conf) {
-  var tag = conf.tag;
-  var swaggerObject = conf.swaggerObject;
-  var propertyName = conf.propertyName;
-
-  // Correct deprecated property.
-  if (propertyName === 'tag') {
-    propertyName = 'tags';
-  }
-
-  if (Array.isArray(tag)) {
-    for (var i = 0; i < tag.length; i = i + 1) {
-      swaggerObject[propertyName].push(tag[i]);
-    }
-  } else {
-    swaggerObject[propertyName].push(tag);
-  }
-}
-
-/**
- * Adds the data in to the swagger object.
- * @function
- * @param {object} swaggerObject - Swagger object which will be written to
- * @param {object[]} data - objects of parsed swagger data from yaml or jsDoc
- *                          comments
- */
-function addDataToSwaggerObject(swaggerObject, data) {
-  for (var i = 0; i < data.length; i = i + 1) {
-    var pathObject = data[i];
-    var propertyNames = Object.getOwnPropertyNames(pathObject);
-    // Iterating the properties of the a given pathObject.
-    for (var j = 0; j < propertyNames.length; j = j + 1) {
-      var propertyName = propertyNames[j];
-      switch (propertyName) {
-        case 'securityDefinition':
-        case 'response':
-        case 'parameter':
-        case 'definition': {
-          var keyName = propertyName + 's';
-          var definitionNames = Object
-            .getOwnPropertyNames(pathObject[propertyName]);
-          for (var k = 0; k < definitionNames.length; k = k + 1) {
-            var definitionName = definitionNames[k];
-            swaggerObject[keyName][definitionName] =
-              pathObject[propertyName][definitionName];
-          }
-          break;
-        }
-        case 'tag':
-        case 'tags': {
-          deprecatedPropertyWarning(propertyName);
-          var tag = pathObject[propertyName];
-          attachTags({
-            tag: tag,
-            swaggerObject: swaggerObject,
-            propertyName: propertyName,
-          });
-          break;
-        }
-        // Assumes a path property if nothing else matches.
-        default: {
-          swaggerObject.paths[propertyName] = objectMerge(
-            swaggerObject.paths[propertyName], pathObject[propertyName]
-          );
-        }
-      }
-    }
-  }
 }
 
 /**
@@ -207,25 +98,15 @@ module.exports = function(options) {
   }
 
   // Build basic swagger json
-  var swaggerObject = [];
-  swaggerObject = options.swaggerDefinition;
-  swaggerObject.swagger = '2.0';
-
-  swaggerObject.paths = swaggerObject.paths || {};
-  swaggerObject.definitions = swaggerObject.definitions || {};
-  swaggerObject.responses = swaggerObject.responses || {};
-  swaggerObject.parameters = swaggerObject.parameters || {};
-  swaggerObject.securityDefinitions = swaggerObject.securityDefinitions || {};
-  swaggerObject.tags = swaggerObject.tags || [];
-
+  var swaggerObject = swaggerHelpers.swaggerizeObj(options.swaggerDefinition);
   var apiPaths = convertGlobPaths(options.apis);
 
   // Parse the documentation in the APIs array.
   for (var i = 0; i < apiPaths.length; i = i + 1) {
     var files = parseApiFile(apiPaths[i]);
     var swaggerJsDocComments = filterJsDocComments(files.jsdoc);
-    addDataToSwaggerObject(swaggerObject, files.yaml);
-    addDataToSwaggerObject(swaggerObject, swaggerJsDocComments);
+    swaggerHelpers.addDataToSwaggerObject(swaggerObject, files.yaml);
+    swaggerHelpers.addDataToSwaggerObject(swaggerObject, swaggerJsDocComments);
   }
 
   parser.parse(swaggerObject, function(err, api) {
@@ -233,5 +114,6 @@ module.exports = function(options) {
       swaggerObject = api;
     }
   });
+
   return swaggerObject;
 };
